% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{diflogadenoma}
\alias{diflogadenoma}
\title{Expression dataset of normal and adenoma samples.}
\format{A data frame with 4175 rows and 13 variables.}
\description{
A dataset containing the colorectal expression data studied by Notterman et al. (2001) and pre processed by von Borries (2008).
}
\references{
D. A. Notterman, U. Alon, A. J. Sierk, and A. J. Levine, Transcriptional gene expression profiles of colorectal adenoma, adenocarcinoma, and normal tissue examined by oligonucleotide arrays, Cancer Research 61 (2001), 3124-3130.
G. von Borries. Partition clustering of High Dimensional Low Sample Size data based on P-Values. PhD Thesis, 2008.
}
\keyword{dataset}
