% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppclust.R
\name{ppclust}
\alias{ppclust}
\title{Clustering Algorithm to HDLSS data.}
\usage{
ppclust(dataset, alpha, ...)
}
\arguments{
\item{dataset}{A numeric matrix or data frame with all numeric columns. If a matrix or data frame, rows correspond to variables (d) and columns correspond to observations (n).}

\item{alpha}{A real number in the range (0, 1) indicanting the threshold parameter to be compared with p-values in the clustering procedure.}

\item{...}{not used.}
}
\value{
Results

A vector of integers indicating the cluster to which each variable is allocated and a cluster cross table of P-values.
}
\description{
Clustering Algorithm to HDLSS data.
}
\examples{
#data(diflogadenoma) # loads data
#cl <- ppclust(diflogadenoma[, 10:13])
#cl
}
