% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppclust_h.R
\name{ppclust_h}
\alias{ppclust_h}
\title{Clustering Algorithm to HDLSS data.}
\usage{
ppclust_h(data, alpha, n.cores = 1, ...)
}
\arguments{
\item{data}{A numeric matrix or data frame with all numeric columns. If a matrix or data frame, rows correspond to variables (d) and columns correspond to observations (n).}

\item{alpha}{A real number in the range (0, 1) indicanting the threshold parameter to be compared with p-values in the clustering procedure.}

\item{n.cores}{A number processor cores (see detectCores).}

\item{...}{not used.}
}
\value{
Results

A vector of integers indicating the cluster to which each variable is allocated and a cluster cross table of P-values.
}
\description{
Clustering Algorithm to HDLSS data.
}
\examples{
#data(diflogadenoma) # loads data
#cl <- ppclust_h(diflogadenoma[, 10:13], n.cores = 2)
#cl
}
